@@ -6,45 +6,50 @@
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"esm-env": "^1.2.2",
 		"@rollup/plugin-dsv": "^3.0.4",
 		"@skeletonlabs/skeleton": "^2.6.0",
 		"@skeletonlabs/tw-plugin": "^0.4.0",
 		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/enhanced-img": "^0.3.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/enhanced-img": "^0.4.4",
+		"@sveltejs/kit": "^2.15.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"@tailwindcss/container-queries": "^0.1.1",
-		"@tailwindcss/forms": "0.5.7",
+		"@tailwindcss/forms": "^0.5.7",
+		"@tsconfig/svelte": "^5.0.0",
 		"autoprefixer": "^10.4.17",
 		"eslint": "^9.1.1",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.30.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"postcss": "^8.4.32",
-		"prettier": "^3.0.0",
-		"prettier-plugin-svelte": "^3.0.0",
-		"svelte": "^4.2.7",
+		"prettier": "^3.1.0",
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte": "^5.17.4",
 		"svelte-check": "^4.0.2",
 		"svelte-virtual-scroll-list": "^1.3.0",
-		"tailwindcss": "^3.3.6",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.0",
-		"vite-plugin-tailwind-purgecss": "^0.3.3"
+		"tailwindcss": "^3.4.17",
+		"tslib": "^2.6.2",
+		"typescript": "^5.5.0",
+		"vite": "^6.0.7",
+		"vite-plugin-tailwind-purgecss": "^0.3.5"
 	},
 	"type": "module",
 	"dependencies": {
 		"@floating-ui/dom": "^1.6.5",
 		"@fortawesome/fontawesome-free": "^6.4.2",
 		"@types/openseadragon": "^3.0.10",
-		"bigger-picture": "^1.1.17",
 		"minisearch": "^7.0.0",
 		"openseadragon": "^5.0.0",
 		"sander": "^0.6.0"
-	}
+	},
+	"description": "This website is a purely static build with a client-side search component. It is hosted by Github and was developed by [Sebastian Flick](https://github.com/flicksolutions), [Peter Dängeli](https://github.com/pdaengeli) and [Hikmat Khan Miakhel](https://github.com/hikmatkhan01) of the [Data Science Lab](https://github.com/dsl-unibe-ch) using Svelte and Svelte-Kit.",
+	"main": "svelte.config.js",
+	"keywords": [],
+	"author": "",
+	"license": "ISC"
 }
