@@ -1,12 +1,9 @@
 <script>
-	let className = '';
-
-	export { className as class };
-	export let dark = false;
+	let { class: className = '', dark = false, children } = $props();
 </script>
 
 <section class="w-full px-8 my-8 {dark ? 'bg-tertiary-500 text-black' : ''}">
 	<div class="container py-4 mx-auto {className}">
-		<slot />
+		{@render children?.()}
 	</div>
 </section>
