@@ -1,7 +1,7 @@
 <script>
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
+
 	import '../app.postcss';
-	import 'bigger-picture/css';
 	import {
 		AppShell,
 		AppBar,
@@ -19,11 +19,10 @@
 	import { storePopup } from '@skeletonlabs/skeleton';
 	import boga from '$lib/assets/BOGA-Logo_Black.svg';
 	import unibe from '$lib/assets/unibe.svg';
-	import { biggerPicture } from '$lib/stores';
-	import BiggerPicture from 'bigger-picture/svelte';
-
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
+	import LightBox from '$lib/components/LightBox.svelte';
 	storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow });
-
 	afterNavigate((/** @type import('@sveltejs/kit').AfterNavigate */ params) => {
 		const isNewPage = params.from?.url?.pathname !== params.to?.url?.pathname;
 		const elemPage = document.querySelector('#page');
@@ -35,8 +34,9 @@
 	initializeStores();
 	const drawerStore = getDrawerStore();
 
-	$: classesActive = (/** @type {string} */ href) =>
-		base + href === $page?.url?.pathname ? 'bg-primary-500' : '';
+	let classesActive = $derived((/** @type {string} */ href) =>
+		base + href === page?.url?.pathname ? 'bg-primary-500' : ''
+	);
 
 	function drawerOpen() {
 		const /** @type {import('@skeletonlabs/skeleton').DrawerSettings} */ s = {
@@ -52,8 +52,8 @@
 		{ slug: 'about us', path: '/about' },
 		{ slug: 'impressum', path: '/impressum' }
 	];
-	let searchtext = '';
-	let otherSearchisVisible = false;
+	let searchtext = $state('');
+	let otherSearchisVisible = $state(false);
 	/**
 	 * @type {IntersectionObserver}
 	 */
@@ -80,11 +80,6 @@
 
 		inputElements.forEach((element) => {
 			observer.observe(element);
-		});
-
-		// initialize lightbox
-		$biggerPicture = BiggerPicture({
-			target: document.body
 		});
 	});
 
@@ -120,7 +115,7 @@
 />
 <!-- App Shell -->
 <AppShell slotPageFooter="bg-surface-200-700-token p-4">
-	<svelte:fragment slot="header">
+	{#snippet header()}
 		<!-- App Bar -->
 		<AppBar padding="px-4" background="bg-surface-100-900-token">
 			<nav class="flex-none items-center h-full hidden md:flex">
@@ -136,26 +131,27 @@
 						<input
 							class="input placeholder-primary-600 ml-2"
 							type="text"
-							placeholder="searchinput..."
+							placeholder="search"
 							bind:value={searchtext}
-							on:change={() => {
+							onchange={() => {
 								const to = searchtext;
 								searchtext = '';
 								goto(`${base}/?s=${to}`);
 							}}
 						/>
 					</label>
-					<a href={`${base}?s=${searchtext}`} class="btn-icon">
+
+					<a href={`${base}?s=${searchtext}`} class="btn-icon" aria-label="Search">
 						<i class="fa-solid fa-search"></i>
 					</a>
 				{/if}
 			</nav>
-			<svelte:fragment slot="lead">
-				<button class="md:!hidden btn-icon" on:click={drawerOpen}>
+			{#snippet lead()}
+				<button class="md:!hidden btn-icon" onclick={drawerOpen} aria-label="Open menu">
 					<i class="fa-solid fa-bars"></i>
 				</button>
-			</svelte:fragment>
-			<svelte:fragment slot="trail">
+			{/snippet}
+			{#snippet trail()}
 				<a href="https://www.unibe.ch" target="_blank" rel="noopener">
 					<img
 						src={unibe}
@@ -176,16 +172,17 @@
 						width="72"
 					/>
 				</a>
-			</svelte:fragment>
+			{/snippet}
 		</AppBar>
-	</svelte:fragment>
+	{/snippet}
 	<!-- Page Route Content -->
-	<slot />
-	<svelte:fragment slot="pageFooter">
+	{@render children?.()}
+	{#snippet pageFooter()}
 		<div class="grid grid-cols-2 lg:ml-10 lg:mr-10 gap-4">
 			<p class="h5 md:h6 lg:h5 col-span-2 justify-self-start">
 				A project of the Herbarium of the Botanical Garden of the University of Bern
 			</p>
 		</div>
-	</svelte:fragment>
+	{/snippet}
 </AppShell>
+<LightBox />
