@@ -1,4 +1,5 @@
-<script>
+<script lang="ts">
+	import { openLightbox } from '$lib/functions';
 	import ContentContainer from '$lib/components/ContentContainer.svelte';
 	import stiftung from '$lib/assets/stiftung_boga.svg';
 	import bgb from '$lib/assets/BGB_Logo_Office_M.svg';
@@ -7,7 +8,6 @@
 	import uws from '$lib/assets/uws.svg';
 	import swisslos from '$lib/assets/swisslos.svg';
 	import fjd from '$lib/assets/fjd.svg';
-	import { openLightbox } from '$lib/functions';
 
 	const logos = [
 		{ src: stiftung, alt: 'Stiftung Boga' },
@@ -34,13 +34,8 @@
 		src="$lib/assets/3E2A0219.jpg"
 		alt="herbarium"
 		class="max-w-full md:max-w-96 md:float-left mx-auto mt-4 md:my-4 md:mr-6 border-8 border-primary-500 anchor cursor-pointer"
-		on:click={(e) => {
-			openLightbox(e.target);
-		}}
-		on:keypress={(e) => {
-			if (e.key === 'Enter' || e.key === ' ') {
-				openLightbox(e.target);
-			}
+		onclick={(e: any) => {
+			openLightbox(e.target.src);
 		}}
 		role="presentation"
 	/>
@@ -82,13 +77,8 @@
 		src="$lib/assets/3E2A0231.jpg"
 		alt="herbarium"
 		class="max-w-full md:max-w-96 md:float-right mx-auto mt-4 md:my-4 md:ml-6 border-8 border-primary-500 anchor cursor-pointer"
-		on:click={(e) => {
-			openLightbox(e.target);
-		}}
-		on:keypress={(e) => {
-			if (e.key === 'Enter' || e.key === ' ') {
-				openLightbox(e.target);
-			}
+		onclick={(e: any) => {
+			openLightbox(e.target.src);
 		}}
 		role="presentation"
 	/>
